+++ tests/test_time.py
@@ -1,10 +1,20 @@
 import unittest
-import typing as tp
 import time
 from satella.time import measure
+from concurrent.futures import Future
 
 
 class TestTime(unittest.TestCase):
+
+    def test_measure_future(self):
+        future = Future()
+        measurement = measure(future)
+        future.set_running_or_notify_cancel()
+        time.sleep(1)
+        future.set_result(2)
+        self.assertEqual(future.result(), 2)
+        self.assertGreater(measurement(), 1)
+
     def test_measure(self):
         with measure() as measurement:
             time.sleep(0.5)